:recycle: (converter) move to_fomatter_string to converter
