+++ src/model.rs
@@ -61,42 +61,6 @@ impl Todo {
         }
     }
 
-    pub fn to_formatted_string(&self) -> String {
-        let mut res: String = "".to_string();
-
-        if self.is_completed {
-            res += "x "
-        }
-
-        if let Some(priority) = &self.priority {
-            res += &("(".to_string() + &priority.to_string() + ") ");
-        }
-
-        if let Some(completion_date) = &self.completion_date {
-            res += &(completion_date.to_string() + " ");
-        }
-
-        if let Some(creation_date) = &self.creation_date {
-            res += &(creation_date.to_string() + " ");
-        }
-
-        res += &self.content;
-
-        if let Some(projects) = &self.projects {
-            for project in projects {
-                res += &(" +".to_string() + project)
-            }
-        }
-
-        if let Some(contexts) = &self.contexts {
-            for context in contexts {
-                res += &(" @".to_string() + context)
-            }
-        }
-
-        res
-    }
-
     pub fn complete(&mut self, completion_date: impl Into<String>) {
         self.is_completed = true;
         self.completion_date = Some(completion_date.into());
@@ -157,123 +121,6 @@ impl Todo {
 #[cfg(test)]
 mod tests {
     use super::Todo;
-
-    mod to_formatted_string {
-        #[test]
-        fn content() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = None;
-            let priority: Option<&str> = None;
-            let projects: Option<String> = None;
-            let contexts: Option<String> = None;
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "content")
-        }
-
-        #[test]
-        fn priority_content() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = None;
-            let priority: Option<&str> = Some("A");
-            let projects: Option<String> = None;
-            let contexts: Option<String> = None;
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "(A) content")
-        }
-
-        #[test]
-        fn creation_date_content() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = Some("2000-1-1");
-            let priority: Option<&str> = None;
-            let projects: Option<String> = None;
-            let contexts: Option<String> = None;
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "2000-1-1 content")
-        }
-
-        #[test]
-        fn priority_creation_date_content() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = Some("2000-1-1");
-            let priority: Option<&str> = Some("A");
-            let projects: Option<String> = None;
-            let contexts: Option<String> = None;
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "(A) 2000-1-1 content")
-        }
-
-        #[test]
-        fn content_projects() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = None;
-            let priority: Option<&str> = None;
-            let projects: Option<String> = Some("projectA,projectB".to_string());
-            let contexts: Option<String> = None;
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "content +projectA +projectB")
-        }
-
-        #[test]
-        fn content_contexts() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = None;
-            let priority: Option<&str> = None;
-            let projects: Option<String> = None;
-            let contexts: Option<String> = Some("contextA,contextB".to_string());
-
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(todo.to_formatted_string(), "content @contextA @contextB")
-        }
-
-        #[test]
-        fn content_projects_contexts() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = None;
-            let priority: Option<&str> = None;
-            let projects: Option<String> = Some("projectA,projectB".to_string());
-            let contexts: Option<String> = Some("contextA,contextB".to_string());
-
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(
-                todo.to_formatted_string(),
-                "content +projectA +projectB @contextA @contextB"
-            )
-        }
-
-        #[test]
-        fn priority_creation_date_content_projects_contexts() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = Some("2000-1-1");
-            let priority: Option<&str> = Some("A");
-            let projects: Option<String> = Some("projectA,projectB".to_string());
-            let contexts: Option<String> = Some("contextA,contextB".to_string());
-
-            let todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            assert_eq!(
-                todo.to_formatted_string(),
-                "(A) 2000-1-1 content +projectA +projectB @contextA @contextB"
-            )
-        }
-
-        #[test]
-        fn complete_priority_creation_date_content_projects_contexts() {
-            let content: &str = "content";
-            let creation_date: Option<&str> = Some("2000-1-1");
-            let priority: Option<&str> = Some("A");
-            let projects: Option<String> = Some("projectA,projectB".to_string());
-            let contexts: Option<String> = Some("contextA,contextB".to_string());
-
-            let mut todo = super::Todo::new(content, creation_date, priority, projects, contexts);
-            todo.complete("2000-1-2");
-
-            assert_eq!(
-                todo.to_formatted_string(),
-                "x (A) 2000-1-2 2000-1-1 content +projectA +projectB @contextA @contextB"
-            )
-        }
-    }
-
     mod complete {
         #[test]
         fn complete() {
